move blog components to blog/ subdir
