+++ website/src/pages/blog.tsx
@@ -1,7 +1,7 @@
 import { graphql } from 'gatsby'
 import * as React from 'react'
-import BlogHeadLinks from '../components/BlogHeadLinks'
-import BlogPosts from '../components/BlogPosts'
+import BlogHeadLinks from '../components/blog/BlogHeadLinks'
+import BlogPosts from '../components/blog/BlogPosts'
 import { ContentPage } from '../components/content/ContentPage'
 import { ContentSection } from '../components/content/ContentSection'
 import Layout from '../components/Layout'
